refactor: Moved Sharings container closer to related files container
