+++ src/drive/ducks/files/index.jsx
@@ -1,6 +1,7 @@
 import React from 'react'
 import Container, { TestFolderView } from './Container'
 export { default as FileOpenerExternal } from './FileOpenerExternal.jsx'
+export { default as SharingsContainer } from './SharingsContainer.jsx'
 
 export const FolderContainer = props => (
   <Container
